+++ src/main/java/babinvas/herokutrialrun/services/SendRequestsService.java
@@ -3,5 +3,5 @@ package babinvas.herokutrialrun.services;
 import babinvas.herokutrialrun.entities.Developer;
 
 public interface SendRequestsService {
-	String receiveDeveloper(long id);
+	String receiveDeveloper(int id);
 }
